+++ Types/EditorTypes/SceneClickHandler.cs
@@ -1,6 +1,6 @@
 ﻿#if UNITY_EDITOR
-#if UNITY_PHYSICS_ENABLED
 using System;
+using JetBrains.Annotations;
 using UnityEngine;
 using UnityEditor;
 
@@ -11,19 +11,33 @@ namespace MyBox.EditorTools
 	/// Set up event, Single Click, Optionally - Cancellable and LayerMask
 	/// Enable handler when needed (or call ToggleState)
 	/// </summary>
+	[PublicAPI]
 	public class SceneClickHandler
 	{
 		/// <param name="onClick">Event called onClick on the scene</param>
+		/// <param name="plane">Plane to read the input</param>
 		/// <param name="singleClick">Single click handler will deactivate itself on click</param>
 		/// <param name="cancellable">Cancellable handler will listen for Escape key to set Enabled to false</param>
-		public SceneClickHandler(Action<Vector3> onClick, bool singleClick = false, bool cancellable = true)
+		public SceneClickHandler(Action<Vector3> onClick, Plane plane, bool singleClick = false, bool cancellable = true)
+			: this(onClick, singleClick, cancellable)
+		{
+			_usePlane = true;
+			_plane = plane;
+		}
+
+		/// <param name="onClick">Event called onClick on the scene</param>
+		/// <param name="singleClick">Single click handler will deactivate itself on click</param>
+		/// <param name="cancellable">Cancellable handler will listen for Escape key to set Enabled to false</param>
+		/// <param name="physics2d">Use Physics2d raycast instead of Physics</param>
+		public SceneClickHandler(Action<Vector3> onClick, bool singleClick = false, bool cancellable = true,
+			bool physics2d = false)
 		{
 			_onClick = onClick;
 
 			SingleClickHandler = singleClick;
+			UsePhysics2D = physics2d;
 			Cancellable = cancellable;
 
-			// TODO: Handle obsolete with ifdefs
 #pragma warning disable 618
 			SceneView.onSceneGUIDelegate += OnSceneGUI;
 #pragma warning restore 618
@@ -39,7 +53,7 @@ namespace MyBox.EditorTools
 
 		public bool Enabled
 		{
-			private get { return _enabled; }
+			private get => _enabled;
 			set
 			{
 				if (value) FocusSceneView();
@@ -58,12 +72,15 @@ namespace MyBox.EditorTools
 		/// <summary>
 		/// Cancellable handler will listen for Escape key to set Enabled to false
 		/// </summary>
-		public bool Cancellable;
+		public bool Cancellable { get; set; }
 
 		/// <summary>
 		/// Single click handler will deactivate itself on click
 		/// </summary>
-		public bool SingleClickHandler;
+		public bool SingleClickHandler { get; set; }
+		
+		public bool UsePhysics2D { get; set; }
+
 
 		public void SetLayerMask(LayerMask mask)
 		{
@@ -71,10 +88,11 @@ namespace MyBox.EditorTools
 			_mask = mask;
 		}
 
-		public void ToggleState()
-		{
-			Enabled = !Enabled;
-		}
+		public void ToggleState() => Enabled = !Enabled;
+		
+		public Color HandleColor { get; set; } = Color.white;
+
+		public float HandleRadius { get; set; } = .3f;
 
 
 		private readonly Action<Vector3> _onClick;
@@ -82,26 +100,69 @@ namespace MyBox.EditorTools
 		private bool _enabled;
 		private bool _useMask;
 		private LayerMask _mask;
+		private bool _usePlane;
+		private Plane _plane;
 
 
 		private void OnSceneGUI(SceneView sceneview)
 		{
 			if (!Enabled) return;
 
-			var ray = HandleUtility.GUIPointToWorldRay(Event.current.mousePosition);
-			RaycastHit hit;
-			if (_useMask ? Physics.Raycast(ray, out hit, _mask.value) : Physics.Raycast(ray, out hit))
+			var point = _usePlane ? GetPlanePoint() : UsePhysics2D ? GetRaycast2dPoint() : GetRaycastPoint();
+
+			if (point.Hit != null)
 			{
 				var color = Handles.color;
-				Handles.color = Color.white;
-				Handles.DrawWireDisc(hit.point, Vector3.up, .3f);
+				Handles.color = HandleColor;
+				Handles.DrawWireDisc(point.Hit.Value, point.Normal, HandleRadius);
 				Handles.color = color;
 
-				if (Handles.Button(Vector3.zero, SceneView.currentDrawingSceneView.rotation, 30, 5000, Handles.RectangleHandleCap))
-					HandleClick(hit.point);
+				if (Handles.Button(Vector3.zero, SceneView.currentDrawingSceneView.rotation, 30, 5000,
+					    Handles.RectangleHandleCap))
+					HandleClick(point.Hit.Value);
 			}
+			
 
-			if (Event.current.type == EventType.KeyDown && Event.current.keyCode == KeyCode.Escape) HandleEscape();
+			if (EscapeInput()) HandleEscape();
+			return;
+
+
+			(Vector3? Hit, Vector3 Normal) GetRaycastPoint()
+			{
+#if UNITY_PHYSICS_ENABLED
+				var ray = HandleUtility.GUIPointToWorldRay(Event.current.mousePosition);
+				if (_useMask ? Physics.Raycast(ray, out var hit, _mask.value) : Physics.Raycast(ray, out hit))
+					return (hit.point, hit.normal);
+#else
+				WarningsPool.LogWarning("SceneClickHandler caused: PHYSICS is not enabled. Use Physics2d or Plane mode instead");
+#endif		
+				return (null, Vector3.zero);
+			}
+			
+			(Vector3? Hit, Vector3 Normal) GetRaycast2dPoint()
+			{
+#if UNITY_PHYSICS2D_ENABLED
+				var ray = HandleUtility.GUIPointToWorldRay(Event.current.mousePosition);
+				var hit = _useMask
+					? Physics2D.Raycast(ray.origin, ray.direction, _mask.value)
+					: Physics2D.Raycast(ray.origin, ray.direction);
+				
+				if (hit.collider != null) return (hit.point, hit.normal);
+#else
+				WarningsPool.LogWarning("SceneClickHandler caused: PHYSICS2D is not enabled. Use Physics or Plane mode instead");
+#endif
+				return (null, Vector3.zero);
+			}
+			
+			(Vector3? Hit, Vector3 Normal) GetPlanePoint()
+			{
+				var ray = HandleUtility.GUIPointToWorldRay(Event.current.mousePosition);
+				if (_plane.Raycast(ray, out var enter)) return (ray.GetPoint(enter), _plane.normal);
+				
+				return (null, Vector3.zero);
+			}
+			
+			bool EscapeInput() => Event.current.type == EventType.KeyDown && Event.current.keyCode == KeyCode.Escape;
 		}
 
 		private void HandleClick(Vector3 point)
@@ -123,5 +184,4 @@ namespace MyBox.EditorTools
 		}
 	}
 }
-#endif
 #endif